editoast: move SpeedSectionCache to infra_cache
