+++ editoast/src/infra_cache/object_cache.rs
@@ -4,6 +4,7 @@ mod electrification_cache;
 mod operational_point_cache;
 mod route_cache;
 mod signal_cache;
+mod speed_section_cache;
 mod switch_cache;
 mod switch_type_cache;
 mod track_section_cache;
